Move WandbLogger to logger.py and apply suggestions from code review
