+++ mmf/trainers/callbacks/logistics.py
@@ -6,12 +6,12 @@ from mmf.trainers.callbacks.base import Callback
 from mmf.utils.configuration import get_mmf_env
 from mmf.utils.logger import (
     TensorboardLogger,
+    WandbLogger,
     calculate_time_left,
     setup_output_folder,
     summarize_report,
 )
 from mmf.utils.timer import Timer
-from mmf.utils.wandb_logger import WandbLogger
 
 
 logger = logging.getLogger(__name__)
